Splitting implementation into several files in anticipation of growth of the code base.
